Fix some issues with delete()
